Move LockedByText to a new file
