+++ src/features/CourtCaseList/tags/LockedByTag/LockedByTag.tsx
@@ -1,30 +1,14 @@
 import ConditionalRender from "components/ConditionalRender"
-import { Tag } from "govuk-react"
 import Image from "next/image"
 import { useState } from "react"
 import { createUseStyles } from "react-jss"
-import { tagBlue, textBlue } from "utils/colours"
 import { useCustomStyles } from "../../../../../styles/customStyles"
+import LockedByText from "./LockedByText"
 
 const useStyles = createUseStyles({
-  LockedByTag: {
-    display: "inline-flex",
-    flexDirection: "row",
-    flexWrap: "nowrap",
-    alignItems: "center",
-    gap: 5
-  },
   LockedIcon: {
     // Change colour from black to GDS text-blue (#144e81) with CSS filters
     filter: "invert(12%) sepia(70%) saturate(4629%) hue-rotate(197deg) brightness(97%) contrast(84%)"
-  },
-  LockedByText: {
-    marginTop: 4,
-    marginBottom: 2,
-    fontWeight: "normal",
-    color: "black",
-    letterSpacing: "0.5px",
-    textTransform: "none"
   },
   LockedByURL: {
     textDecoration: "underline",
@@ -114,29 +98,6 @@ const LockedByButton = ({
         />
       )}
     </>
-  )
-}
-
-interface LockedByTextProps {
-  lockedBy?: string | null
-  unlockPath?: string
-}
-
-const LockedByText = ({ lockedBy, unlockPath }: LockedByTextProps) => {
-  const classes = useStyles()
-  return (
-    <Tag backgroundColor={tagBlue} color={textBlue} className={`locked-by-tag ${classes.LockedByTag}`}>
-      <div className={classes.LockedByTag}>
-        <Image
-          src={"/bichard/assets/images/lock.svg"}
-          width={18}
-          height={18}
-          className={unlockPath ? classes.LockedIcon : undefined}
-          alt="Lock icon"
-        />
-        <span className={classes.LockedByText}>{lockedBy}</span>
-      </div>
-    </Tag>
   )
 }
